Create CollisionPreventionDetection.py
This builds on the testing.py file. You can move the two cars around using WASD and arrow keys. When you get close to a collision, it will warn you in terminal and tell you in which direction you'll hit. Next step is to get this working in the main simulation with the biggest challenge being how to deal with a changing heading (the heading in this simulation is always upward). Also, the sensitivity to a collision can be adjusted with some of the constants at the top of the program.